requirements.txt
accelerate==1.12.0
aiofile==3.9.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
ansible==13.4.0
ansible-core==2.20.3
anyio==4.12.1
apache-libcloud==3.9.0
attrs==25.4.0
audioop-lts==0.2.2
Authlib==1.6.8
autopage==0.6.0
av==16.1.0
bcrypt==5.0.0
beartype==0.22.9
beautifulsoup4==4.14.3
blinker==1.9.0
boto3==1.42.59
botocore==1.42.59
brotli==1.2.0
bs4==0.0.2
cachetools==6.2.6
caio==0.9.25
cattrs==26.1.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.4
cheroot==11.1.2
click==8.3.1
cliff==4.13.2
cmd2==3.2.2
configparser==7.2.0
cryptography==46.0.5
ctranslate2==4.7.1
cyclopts==4.6.0
ddgs==9.10.0
defusedxml==0.7.1
deprecation==2.1.0
diffusers==0.36.0
dnspython==2.8.0
docker==7.1.0
docstring_parser==0.17.0
docutils==0.22.4
duckduckgo_search==8.1.1
easyocr==1.7.2
edge-tts==7.2.7
email-validator==2.3.0
evdev==1.9.3
exceptiongroup==1.3.1
fake-useragent==2.2.0
fastapi==0.135.1
faster-whisper==1.2.1
fastmcp==3.0.2
filelock==3.20.0
Flask==3.1.3
flask-cors==6.0.2
flatbuffers==25.12.19
frozenlist==1.8.0
fsspec==2025.12.0
gitdb==4.0.12
GitPython==3.1.46
google-api-core==2.30.0
google-api-python-client==2.190.0
google-auth==2.48.0
google-auth-httplib2==0.3.0
google-auth-oauthlib==1.2.4
googleapis-common-protos==1.72.0
greenlet==3.3.2
h11==0.16.0
h2==4.3.0
hf-xet==1.3.1
hpack==4.1.0
httpcore==1.0.9
httplib2==0.31.2
httpx==0.28.1
httpx-sse==0.4.3
huggingface_hub==1.5.0
hvac==2.4.0
hyperframe==6.1.0
idna==3.11
im==1.19.1
ImageIO==2.37.2
importlib_metadata==8.7.1
invoke==2.2.1
itsdangerous==2.2.0
jaraco.classes==3.4.0
jaraco.context==6.1.0
jaraco.functools==4.4.0
jeepney==0.9.0
Jinja2==3.1.6
jmespath==1.1.0
jsonref==1.1.0
jsonschema==4.26.0
jsonschema-path==0.4.4
jsonschema-specifications==2025.9.1
keyring==25.7.0
lazy_loader==0.4
linkify-it-py==2.0.3
llvmlite==0.46.0
lxml==6.0.2
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mcp==1.26.0
mcp-server-git==2026.1.14
mdit-py-plugins==0.5.0
mdurl==0.1.2
mmh3==5.2.0
more-itertools==10.8.0
MouseInfo==0.1.3
mpmath==1.3.0
multidict==6.7.1
netaddr==1.3.0
networkx==3.6.1
ninja==1.13.0
numba==0.64.0
numpy==2.3.5
oauthlib==3.3.1
ollama==0.6.1
onnxruntime==1.24.2
openai-whisper==20250625
openapi-pydantic==0.5.1
opencv-python-headless==4.13.0.92
opentelemetry-api==1.39.1
packaging==26.0
paramiko==4.0.0
pathable==0.5.0
pbr==7.0.3
pillow==12.0.0
piper-tts==1.4.1
platformdirs==4.9.2
playsound3==3.3.1
playwright==1.58.0
ply==3.11
postgrest==2.28.0
prettytable==3.17.0
primp==1.1.1
propcache==0.4.1
proto-plus==1.27.1
protobuf==6.33.5
psutil==7.2.2
ptyprocess==0.7.0
py-key-value-aio==0.4.4
pyasn1==0.6.2
pyasn1_modules==0.4.2
PyAudio==0.2.14
PyAutoGUI==0.9.54
pyclipper==1.4.0
pycparser==3.0
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic_core==2.41.5
pyee==13.0.1
pygame==2.6.1
PyGetWindow==0.0.9
Pygments==2.19.2
pyiceberg==0.11.0
PyJWT==2.11.0
PyMsgBox==2.0.1
PyNaCl==1.6.2
pynput==1.8.1
pyparsing==3.3.2
pyperclip==1.11.0
PyRect==0.2.0
pyroaring==1.0.3
PyScreeze==1.0.1
python-bidi==0.6.7
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.22
python-xlib==0.33
python3-xlib==0.15
pytweening==1.2.0
PyYAML==6.0.3
qwen-vl-utils==0.0.14
radl==1.3.5
realtime==2.28.0
redis==7.2.1
referencing==0.37.0
regex==2026.2.19
requests==2.32.5
requests-cache==1.3.0
requests-oauthlib==2.0.0
resolvelib==1.2.1
rich==14.3.3
rich-argparse==1.7.2
rich-rst==1.3.2
rpds-py==0.30.0
rsa==4.9.1
s3transfer==0.16.0
safetensors==0.7.0
scar==4.3.1
scikit-image==0.26.0
scipy==1.17.1
scp==0.15.0
SecretStorage==3.5.0
sentencepiece==0.2.1
setuptools==70.2.0
shapely==2.1.2
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
socksio==1.0.0
soupsieve==2.8.3
SpeechRecognition==3.14.5
sse-starlette==3.2.0
standard-aifc==3.13.0
standard-chunk==3.13.0
starlette==0.52.1
stevedore==5.7.0
storage3==2.28.0
StrEnum==0.4.15
strictyaml==1.7.3
suds-community==1.2.0
supabase==2.28.0
supabase-auth==2.28.0
supabase-functions==2.28.0
sympy==1.14.0
tabulate==0.9.0
tenacity==9.1.4
textual==8.0.0
tifffile==2026.2.24
tiktoken==0.12.0
tokenizers==0.22.2
torch==2.10.0+cpu
torchaudio==2.10.0+cpu
torchvision==0.25.0+cpu
tosca-parser==2.14.0
tqdm==4.67.3
transformers==5.2.0
triton==3.6.0
typer==0.24.1
typer-slim==0.24.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uc-micro-py==1.0.3
uritemplate==4.2.0
url-normalize==2.2.1
urllib3==2.6.3
uvicorn==0.41.0
watchfiles==1.1.1
wcwidth==0.6.0
websockets==15.0.1
Werkzeug==3.1.6
xmltodict==1.0.4
yarl==1.22.0
youtube-dl==2021.12.17
youtube-transcript-api==1.2.4
YouTubeTranscript==1.0
zipp==3.23.0
zstandard==0.25.0
